Add hover effect and content update to portfolio
Updated the CSS to add a hover effect on images and made content modifications in the index.html. The hover effect will help make the portfolio images pop. Image hyperlinks were also removed and replaced with direct images for a cleaner feel. A Github link was added in navigation and placeholder images were removed. Modifications made to the html file ensure relevant, up-to-date information.